requirement.txt
asgiref==3.6.0
attrs==22.2.0
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
dj-mongo-backend==0.1.4
Django==4.1.7
django-cors-headers==3.14.0
django-rest-framework-mongoengine==3.4.1
django-rest-registration==0.7.3
djangorestframework==3.12.2
drf-spectacular==0.26.1
exceptiongroup==1.1.1
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
mongoengine==0.27.0
mypy-extensions==1.0.0
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pathspec==0.11.1
platformdirs==3.1.1
pluggy==1.0.0
pymongo==3.13.0
pyrsistent==0.19.3
pytest==7.2.2
python-dateutil==2.8.2
pytz==2020.5
PyYAML==6.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.15
